requirements.txt
Flask==0.12.0
pymongo==3.6.0
python-jose==2.0.1
nose==1.3.7